replace unique skill bonuses with resistances
Instead of an Air spell being cast with Charisma + Seafaring, it now
casts with Charisma + Air at all times.

The \spell function now uses the fifth argument to show what kinds of
thing resist the spell.

If the spell is resisted by a person, you can enter
\roll{Strength}{Athletics} to show it being resisted by 'Strength +
Athletics'.  Otherwise, input 'the wind speed', and this will read 'the
TN depends on the wind speed'.