recalculate cycles
Cycles should have a simple, convenient way to calculate, so the Judge
can look at a Mundane calendar, and easily see the date in Fenestra.

This calendar divides the year into three parts, and asks if this is the
first, second, or third month in the quarter.  The cycle equals 'month x
2' (so February = 4) and the first half of the month gets a -1.  So
February 1st occurs during the fifth cycle, i.e. 'Sylfs'.

The day in Fenestra simply equals our day x 4.  After the 15th of each
month, we subtract 15, and start again (so April 16th = cycle 2, day 4).