parallel compilations!
Projects had to be compiled independently, otherwise the changes to
their shared output directory would conflict.  But now, instead of
listing the page-numbers in the core book by putting everything in
'../rubbish', the projects place all outputs in 'rubbish/', (e.g.
rubbish/core.aux) and then the target 'ref' copies that aux file to
../aux_share/.

All projects check the shared aux directory for interesting files, so
the 'oneshot' project can check for:

../aux_share/core.aux
../aux_share/judgement.aux

The files are then put into the local 'rubbish/' directory.

HOWEVER, latexmk will not check if these files change, because it has a
list of dependencies, and those files are not on the list (because they
did not exist when latexmk listed the dependencies).  So the layout.sty
file now says to always check for those files, even if they don't exist,
which brings further warning messages.

This is fine, because the various layers of TeX already produce so many
warning messages that nobody could possibly read them all, so we don't.