replace glossary `short={}` entries with `long={}`
Using the `short=` property of glossary entires (like `weight`) meant
that this short definition was printed instead of the word, so writing
`\gls{weight}` showed the short definition, not the word 'Weight'.

Using the `long={},` property doesn't seem to interfere with anything,
though it's unfortuante that we'll be using `\glsentrylong{weight}` to
refer to a glossary entry's shorter description.