give 'warden' short form 'ward'
We understand 'yes, my lord', and 'no, my lord' as dice-flavoured 'Dark
Ages' deference.  But it makes no sense in Fenestra, where nobody has
the title 'lord'.  However, they have a pseudo-noble position called
'warden'.  By adding the short-form, characters can say 'yes, my ward',
and so on, and it's so close that readers and players may auto-correct
it the first few times they hear it.