restore original CR formula
The CR formula was based on the `damage` counter, which once took '4' as
the basic score.  The standard Damage is 1D6, which represents 4 Damage
(just as 1D6+4 Damage = 2D6 Damage).  The `damage` counter was
simplified to take 0 as 1D6, so it could take bonuses easier, but for
the CR rating, we just add 4 to return to the original formula.