create recfile template
People can't write for BIND without a bunch of LaTeX.  This is fine, as
they can use a pen.  However, the statblocks are *not* fine.  Writing
them with a pen sounds punishing.  The recfile template should be much
easier.

So far it looks like this:

    Name: auroch
    Strength: 2
    Dexterity: -2
    Speed: 1
    Wits: 1
    Abilities: \quadruped

The statblocks will get other stats soon (like Skills).  Hopefully this
will make the statblocks more accessible.