remove order dependency from skills
Using `\edef` resulted in abilities which responded to *current* weight.
If another ability added to weight later, the first ability would not
take account of the change.

The new method simply uses `\appto{\Abilities}` to set counters and
commands, which do nothing until printed.  Anything written outside this
is processed in order, as usual.